pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "galleguimetro"
version = "0.2.0"
description = "Sistema de análisis de opciones financieras en tiempo real"
authors = [
    { name = "Federico", email = "dev@galleguimetro.app" }
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "websockets>=12.0",
    "httpx>=0.25.0",
    "quantlib>=1.26",
    "python-dotenv>=1.0.0",
    "SQLAlchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "slowapi>=0.1.9",
]
requires-python = ">=3.11"
readme = "README.md"
license = {file = "LICENSE"}
keywords = ["finance", "options", "greeks", "trading", "analysis", "fastapi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]

[project.optional-dependencies]
dde = ["pywin32>=306"]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "alembic>=1.12.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.urls]
Homepage = "https://github.com/tanobuffone/galleguimetro"
Repository = "https://github.com/tanobuffone/galleguimetro.git"

[tool.hatch.build.targets.wheel]
packages = ["galleguimetro"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
